% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_functions.R
\name{f1}
\alias{f1}
\title{does something}
\usage{
f1(use_f2 = FALSE)
}
\arguments{
\item{use_f2}{should \code{f2} be called or not}
}
\value{
nothing
}
\description{
this function calls \code{f2}, some of the time
}

